+++ refinedstorage-platform-common/src/main/java/com/refinedmods/refinedstorage/platform/common/grid/screen/hint/ItemGridInsertionHint.java
@@ -2,7 +2,7 @@ package com.refinedmods.refinedstorage.platform.common.grid.screen.hint;
 
 import com.refinedmods.refinedstorage.platform.api.grid.GridInsertionHint;
 import com.refinedmods.refinedstorage.platform.api.support.AmountFormatting;
-import com.refinedmods.refinedstorage.platform.common.support.tooltip.MouseWithIconClientTooltipComponent;
+import com.refinedmods.refinedstorage.platform.common.support.tooltip.MouseClientTooltipComponent;
 
 import java.util.Optional;
 
@@ -12,9 +12,9 @@ import net.minecraft.world.item.ItemStack;
 public class ItemGridInsertionHint implements GridInsertionHint {
     @Override
     public Optional<ClientTooltipComponent> getHint(final ItemStack carried) {
-        return Optional.of(new MouseWithIconClientTooltipComponent(
-            MouseWithIconClientTooltipComponent.Type.LEFT,
-            (graphics, x, y) -> graphics.renderItem(carried, x, y),
+        return Optional.of(MouseClientTooltipComponent.item(
+            MouseClientTooltipComponent.Type.LEFT,
+            carried,
             carried.getCount() == 1 ? null : AmountFormatting.format(carried.getCount())
         ));
     }