+++ refinedstorage-platform-common/src/main/java/com/refinedmods/refinedstorage/platform/common/grid/screen/hint/SingleItemGridInsertionHint.java
@@ -1,7 +1,7 @@
 package com.refinedmods.refinedstorage.platform.common.grid.screen.hint;
 
 import com.refinedmods.refinedstorage.platform.api.grid.GridInsertionHint;
-import com.refinedmods.refinedstorage.platform.common.support.tooltip.MouseWithIconClientTooltipComponent;
+import com.refinedmods.refinedstorage.platform.common.support.tooltip.MouseClientTooltipComponent;
 
 import java.util.Optional;
 
@@ -11,9 +11,9 @@ import net.minecraft.world.item.ItemStack;
 public class SingleItemGridInsertionHint implements GridInsertionHint {
     @Override
     public Optional<ClientTooltipComponent> getHint(final ItemStack carried) {
-        return Optional.of(new MouseWithIconClientTooltipComponent(
-            MouseWithIconClientTooltipComponent.Type.RIGHT,
-            (graphics, x, y) -> graphics.renderItem(carried, x, y),
+        return Optional.of(MouseClientTooltipComponent.item(
+            MouseClientTooltipComponent.Type.RIGHT,
+            carried,
             null
         ));
     }